Implement controller plugin

+++ src/messages.rs
@@ -3,8 +3,8 @@ use nphysics2d::algebra::Velocity2;
 
 #[derive(Copy, Clone, Default, Debug)]
 pub struct Point {
-    pub x: f64,
-    pub y: f64
+    pub x: f32,
+    pub y: f32
 }
 
 impl Into<[f32; 2]> for Point {
@@ -14,23 +14,25 @@ impl Into<[f32; 2]> for Point {
 }
 
 #[derive(Clone, Default, Debug)]
+#[repr(C)]
 pub struct PointCloud {
     pub points: Vec<Point>
 }
 
+#[repr(C)]
 #[derive(Copy, Clone, Default, Debug)]
 pub struct Pose {
-    pub x: f64,
-    pub y: f64,
-    pub theta: f64
+    pub x: f32,
+    pub y: f32,
+    pub theta: f32
 }
 
 impl From<Isometry2<f64>> for Pose {
     fn from(i: Isometry2<f64>) -> Self {
         Pose {
-            x: i.translation.vector[0],
-            y: i.translation.vector[1],
-            theta: i.rotation.angle()
+            x: i.translation.vector[0] as f32,
+            y: i.translation.vector[1] as f32,
+            theta: i.rotation.angle() as f32
         }
     }
 }
